requirements.txt
transformers==4.56.2
torch==2.8.0
sentencepiece==0.2.1
llama_cpp_python==0.3.16
fastapi==0.119.0
uvicorn==0.37.0
requests==2.32.5
pydantic==2.12.2